JS:s animate.js 분리
