+++ js/index.js
@@ -1,6 +1,6 @@
 import { includeHTML } from "./include-html.js";
-import { animate } from "./animate.js";
+import { main } from "./main.js";
 
 includeHTML().then(async () => {
-  animate();
+  main();
 });
